pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "CommunityBuilder.Models"
version = "0.1.0"
description = "Shared Django models for CommunityBuilder API"
requires-python = ">=3.8"
dependencies = [
    "Django>=4.0",
    "djangorestframework",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app_models*"]
